scaled data, balanced classifier, correlated data, and analyzed features with most prominent corr compared to class
Scaled the amount data to make it similar to the other features. Created a new dataset that balances both classifiers for a better analysis of what features increase the chances of a fraudulent charge. Correlated both source data and balanced data to visualize the difference. Analyzed features with the most correlation to the Classifier (Class). Started section for  determining anomaly handling which will be based on further data analysis.